+++ CHANGELOG.md
@@ -1,2 +1,7 @@
+* **2018-08-24:**
+    * Add tzdata package to fix localtime/TZ issue
+    * Update docker-entrypoint.sh, change LOGDIR to LOGFILE
+    * Update build hook script
+---
 * **2018-08-19:**
     * Initial Dockerfile, script, etc. creation.